fix crash on deleting control point while dragging
