+++ toonz/sources/tnztools/controlpointeditortool.cpp
@@ -1,11 +1,9 @@
-
+#include "controlpointeditortool.h"
 
 #include "tundo.h"
 #include "tthreadmessage.h"
 #include "tvectorimage.h"
 #include "drawutil.h"
-#include "controlpointselection.h"
-#include "tproperty.h"
 #include "tenv.h"
 
 #include "tools/tool.h"
@@ -22,8 +20,6 @@
 #include "toonz/stage2.h"
 #include "toonz/tstageobject.h"
 
-// For Qt translation support
-#include <QCoreApplication>
 #include <QKeyEvent>
 
 using namespace ToolUtils;
@@ -121,126 +117,6 @@ void getSegmentParameter(ControlPointEditorStroke *cpEditor, int beforeIndex,
 }
 
 }  // namespace
-//=============================================================================
-// ControlPointEditorTool
-//-----------------------------------------------------------------------------
-
-class ControlPointEditorTool final : public TTool {
-  Q_DECLARE_TR_FUNCTIONS(ControlPointEditorTool)
-
-  bool m_draw;
-  bool m_isMenuViewed;
-  int m_lastPointSelected;
-  bool m_isImageChanged;
-  ControlPointSelection m_selection;
-  ControlPointEditorStroke m_controlPointEditorStroke;
-  std::pair<int, int> m_moveSegmentLimitation;  // Indici dei punti di controllo
-                                                // che limitano la curva da
-                                                // muovere
-  ControlPointEditorStroke m_moveControlPointEditorStroke;  // Usate per muovere
-                                                            // la curva durante
-                                                            // il drag.
-  TRectD m_selectingRect;
-  TPointD m_pos;
-
-  TPropertyGroup m_prop;
-  TEnumProperty m_selectType;
-
-  TBoolProperty
-      m_autoSelectDrawing;  // Consente di scegliere se swichare tra i livelli.
-
-  TBoolProperty m_snap;
-  TEnumProperty m_snapSensitivity;
-  double m_snapMinDistance;
-  bool m_foundSnap;
-  TPointD m_snapPoint;
-
-  TPointD m_firstPos;       // The first point inserted in m_track
-  StrokeGenerator m_track;  // Lazo selection generator.
-  TStroke *m_stroke;        // Stores the stroke generated by m_track.
-
-  enum Action {
-    NONE,
-    RECT_SELECTION,
-    FREEHAND_SELECTION,
-    CP_MOVEMENT,
-    SEGMENT_MOVEMENT,
-    IN_SPEED_MOVEMENT,
-    OUT_SPEED_MOVEMENT
-  };
-  Action m_action;
-
-  enum CursorType { NORMAL, ADD, EDIT_SPEED, EDIT_SEGMENT, NO_ACTIVE };
-  CursorType m_cursorType;
-
-  TUndo *m_undo;
-
-  void selectRegion(TStroke *stroke);
-  void startFreehand(const TPointD &pos);
-  void freehandDrag(const TPointD &pos);
-  void closeFreehand(const TPointD &pos);
-
-public:
-  ControlPointEditorTool();
-
-  ToolType getToolType() const override { return TTool::LevelWriteTool; }
-
-  void updateTranslation() override;
-
-  TPropertyGroup *getProperties(int targetType) override { return &m_prop; }
-
-  // da TSelectionOwner: chiamato quando la selezione corrente viene cambiata
-  void onSelectionChanged() { invalidate(); }
-
-  // da TSelectionOwner: chiamato quando si vuole ripristinare una vecchia
-  // selezione
-  // attualmente non usato
-  bool select(const TSelection *) { return false; }
-  ControlPointEditorStroke getControlPointEditorStroke() {
-    return m_controlPointEditorStroke;
-  };
-
-  void initUndo();
-
-  void getNearestStrokeColumnIndexes(std::vector<int> &indexes, TPointD pos);
-
-  void drawMovingSegment();
-  void drawControlPoint();
-  void draw() override;
-  void mouseMove(const TPointD &pos, const TMouseEvent &e) override;
-  void leftButtonDown(const TPointD &pos, const TMouseEvent &e) override;
-  void rightButtonDown(const TPointD &pos, const TMouseEvent &) override;
-
-  void moveControlPoints(const TPointD &delta);
-  void moveSpeed(const TPointD &delta, bool isIn);
-  void moveSegment(const TPointD &delta, bool dragging, bool isShiftPressed);
-
-  void leftButtonDrag(const TPointD &pos, const TMouseEvent &e) override;
-  void leftButtonUp(const TPointD &pos, const TMouseEvent &e) override;
-  void addContextMenuItems(QMenu *menu) override;
-
-  void linkSpeedInOut(int index);
-  void unlinkSpeedInOut(int pointIndex);
-
-  bool keyDown(QKeyEvent *event) override;
-  void onEnter() override;
-  void onLeave() override;
-  bool onPropertyChanged(std::string propertyName) override;
-
-  void onActivate() override;
-  void onDeactivate() override;
-  void onImageChanged() override;
-  int getCursorId() const override;
-
-  // returns true if the pressed key is recognized and processed.
-  bool isEventAcceptable(QEvent *e) override;
-
-  TPointD calculateSnap(TPointD pos);
-  void drawSnap();
-  TPointD getSnap(TPointD pos);
-  void resetSnap();
-
-} controlPointEditorTool;
 
 //-----------------------------------------------------------------------------
 
@@ -560,7 +436,6 @@ void ControlPointEditorTool::leftButtonDown(const TPointD &pos,
     getViewer()->doPickGuideStroke(pos);
     return;
   }
-  m_selection.setArePointsDeleted(false);
   m_pos           = pos;
   double pix      = getPixelSize() * 2.0f;
   double maxDist  = 5 * pix;
@@ -796,13 +671,6 @@ void ControlPointEditorTool::moveSegment(const TPointD &delta, bool dragging,
 
 void ControlPointEditorTool::leftButtonDrag(const TPointD &pos,
                                             const TMouseEvent &e) {
-  if (m_selection.getPointsDeleted()) {
-    m_selection.selectNone();
-    m_lastPointSelected = -1;
-    if (m_undo) delete (m_undo);
-    m_undo = 0;
-    return;
-  }
   TVectorImageP vi(getImage(true));
   int currentStroke = m_controlPointEditorStroke.getStrokeIndex();
   if (!vi || currentStroke == -1 || m_action == NONE) return;
@@ -820,15 +688,6 @@ void ControlPointEditorTool::leftButtonDrag(const TPointD &pos,
       TThickPoint cp;
       TPointD controlPoint;
       TPointD newPos;
-      int count = m_controlPointEditorStroke.getControlPointCount();
-      if (m_lastPointSelected > count - 1) {
-        m_selection.selectNone();
-        m_lastPointSelected = -1;
-        if (m_undo) delete (m_undo);
-        m_undo = 0;
-
-        return;
-      }
       cp = m_controlPointEditorStroke.getControlPoint(m_lastPointSelected);
       controlPoint = TPointD(cp.x, cp.y);
       newPos       = calculateSnap(pos);
@@ -895,12 +754,6 @@ void ControlPointEditorTool::selectRegion(TStroke *stroke) {
 
 void ControlPointEditorTool::leftButtonUp(const TPointD &realPos,
                                           const TMouseEvent &e) {
-  if (m_selection.getPointsDeleted()) {
-    m_selection.setArePointsDeleted(false);
-    if (m_undo) delete (m_undo);
-    m_undo = 0;
-    return;
-  }
   TVectorImageP vi(getImage(true));
   int currentStroke = m_controlPointEditorStroke.getStrokeIndex();
   if (!vi || currentStroke == -1) return;
@@ -937,7 +790,6 @@ void ControlPointEditorTool::leftButtonUp(const TPointD &realPos,
       m_isImageChanged = false;
     }
   }
-  std::set<int> oldPoints = m_selection.getSelectedPoints();
 
   if (m_action == NONE || !m_isImageChanged) {
     m_undo = 0;
@@ -947,16 +799,6 @@ void ControlPointEditorTool::leftButtonUp(const TPointD &realPos,
 
   notifyImageChanged();
   invalidate();
-
-  if (m_action == CP_MOVEMENT) {
-    if (oldPoints.size() >= 1) {
-      for (auto point : oldPoints) {
-        m_selection.select(point);
-        m_lastPointSelected = point;
-      }
-      m_selection.makeCurrent();
-    }
-  }
 
   // Registro l'UNDO
   if (m_undo) {
@@ -998,9 +840,6 @@ bool ControlPointEditorTool::keyDown(QKeyEvent *event) {
   TVectorImageP vi(getImage(true));
   if (!vi || (vi && m_selection.isEmpty())) return false;
 
-  // Inizializzo l'UNDO
-  initUndo();
-
   TPointD delta;
   switch (event->key()) {
   case Qt::Key_Up:
@@ -1019,6 +858,9 @@ bool ControlPointEditorTool::keyDown(QKeyEvent *event) {
     return false;
     break;
   }
+  // Inizializzo l'UNDO
+  initUndo();
+
   moveControlPoints(delta);
 
   invalidate();
@@ -1116,8 +958,8 @@ void ControlPointEditorTool::onImageChanged() {
     m_controlPointEditorStroke.setStroke((TVectorImage *)0, -1);
     return;
   } else {
-    m_selection.selectNone();
-    m_controlPointEditorStroke.setStroke(vi, currentStroke);
+    if (m_controlPointEditorStroke.setStroke(vi, currentStroke))
+      m_selection.selectNone();
   }
 }
 
@@ -1190,3 +1032,5 @@ void ControlPointEditorTool::closeFreehand(const TPointD &pos) {
   m_stroke     = m_track.makeStroke(error);
   m_stroke->setStyle(1);
 }
+
+ControlPointEditorTool controlPointEditorTool;
